move storagetank_mixed.py into cea folder
